% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbaPlayerTable.R
\name{player_table}
\alias{player_table}
\title{NBA Player List Generator}
\usage{
player_table(year, team = "ALL")
}
\arguments{
\item{year}{the year in which the second half of the season takes place}

\item{team}{(optional) A team's acronym to return a specific team's player. Default is all teams, or you may select one specific team}
}
\value{
a table
}
\description{
NBA Player List Generator
}
